this is a quick demo of a 3d graph
The files in the plot folder were copied from a bolt.new project
I don't like the camel case file names, but that is what bolt used so we can stick with them for now.

In order to no conflict with other changes, this just updates the index.tsx to point at the ScatterPlot instead of the App.

I did also update the App so it wouldn't throw a warning when running outside of CODAP which might help with testing.

The bolt code used tailwind for the css. I did not bring that over so the styling and sizing of things is off.