update eslint and webpack
This followed some of the pattern from the CODAPv3 eslint upgrade. However I tried to not use the compat feature now that most of the plugins support flat files. Also this is using tseslint's `config` function which adds some type checking to the config.

The version of 'globals' had to be overridden because babel/traverse used by cypress/code-coverage used an older version.

Also I updated webpack dev server to the latest version.