fix report upload, always record trace
The report was broken because the action was run on pull which runs against a merge commit with a ref of `refs/pull/25/merge`. The s3-deploy-path doesn't handle this case.

Also the location of the upload would conflict with the uploaded files from the s3-deploy-action, so if the s3-deploy-action happened second it would delete the report.

Finally we want to record the trace always at least to demonstrate traces, and to make it easier to review PRs.